LayoutTests/http/wpt/web-animations/timing-model/animations/set-the-target-effect-of-an-animation-expected.txt


FAIL If new effect is null and old effect is not null, we reset the pending tasks and ready promise is rejected assert_equals: expected "pending" but got "running"
FAIL If animation has a pending pause task, reschedule that task to run as soon as animation is ready. assert_equals: expected "pending" but got "paused"
FAIL If animation has a pending play task, reschedule that task to run as soon as animation is ready to play new effect. assert_equals: expected "pending" but got "finished"
PASS When setting the effect of an animation to the effect of an existing animation, the existing animation's target effect should be set to null. 
PASS After setting the target effect of animation to the target effect of an existing animation, the target effect's timing is updated to reflect the current time of the new animation. 
